pyproject.toml
[project]
name = "rag-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ace-tools>=0.0",
    "beautifulsoup4>=4.13.3",
    "faiss-cpu>=1.10.0",
    "google-generativeai>=0.8.4",
    "langchain-community>=0.3.21",
    "langchain-huggingface>=0.1.2",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
]
